Reorg pallet
